% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celsius-a-Fahrenheit.R
\name{convertir_cf}
\alias{convertir_cf}
\title{Funcion para convertir grados celsius a grados fahrenheit}
\usage{
convertir_cf(temp_celsius)
}
\arguments{
\item{temp_centigrados}{Una temperatura en grados celsius (un vector numerico)}
}
\value{
Un vector numerico en grados fahrenheit
}
\description{
La funcion \code{convertir_cf()} toma una temperatura en grados celsius y la convierte a grados fahrenheit
}
\details{
Estos son detalles extra de la funcion
}
\examples{

convertir_cf(25)
convertir_cf(10)

}
\seealso{
\code{\link[=convertir_fc]{convertir_fc()}}
\code{\link[=clasificar_temperatura]{clasificar_temperatura()}}
}
