% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcion Fahrenheit a Celsius.R
\name{convertir_fc}
\alias{convertir_fc}
\title{De fahrenheit a celsius}
\usage{
convertir_fc(temp_fahrenheit)
}
\arguments{
\item{temp_fahrenheit}{un vector numerico}
}
\value{
Un vector numérico en grados Celsius
}
\description{
La funcion \code{convertir_fc()} pasa de grados fahrenheit a celsius
}
\examples{
convertir_fc(c(0, 32, 100))


}
\seealso{
\code{\link[=convertir_cf]{convertir_cf()}}
}
