% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_datos.R
\name{descargar_datos}
\alias{descargar_datos}
\title{Descargar datos de distintas estaciones meteorologicas}
\usage{
descargar_datos(id_estacion)
}
\arguments{
\item{id_estacion}{un string de letras y numeros}
}
\value{
La informacion que contiene el dataset que fue descargado
}
\description{
La funcion 'descargar_datos()' toma como parametro el id de alguna de las
estaciones que se solicite para luego realizar la descarga y lectura del dataset
}
\details{
Estos son detalles extra de la funcion
}
\examples{
data <- descargar_datos("NH0046")

}
