% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grafico.R
\name{grafico_temperatura}
\alias{grafico_temperatura}
\title{Gráfico de temperatura media mensual para estaciones meteorológicas}
\usage{
grafico_temperatura(dataset, colores = NULL, titulo = "Temperatura Abrigo")
}
\arguments{
\item{dataset}{Un data frame que contiene la información de las estaciones meteorológicas. Debe tener al menos las columnas \code{fecha} (tipo Date o POSIXt), \code{temperatura_abrigo_150cm} y \code{id} (identificador de estación).}

\item{colores}{Un vector opcional de colores a usar para las distintas estaciones en el gráfico. Si no se especifica o si no contiene suficientes colores, se generará automáticamente un conjunto de colores aleatorios.}

\item{titulo}{Un string con el título del gráfico. El valor predeterminado es \code{"Temperatura Abrigo"}.}
}
\value{
Un objeto \code{ggplot} que representa un gráfico de línea de la temperatura media mensual
}
\description{
La funcion \code{grafico_temperatura} toma un dataset con informacion sobre estaciones meteorologicas y devuelve un grafico de la temperatura media mensual por estación.
}
\details{
Estos son detalles extra de la funcion
}
\examples{
grafico_temperatura(NH0046)

}
