% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tablas.R
\name{tabla_resumen_temperatura}
\alias{tabla_resumen_temperatura}
\title{Tabla de resumen}
\usage{
tabla_resumen_temperatura(nombre_archivo)
}
\arguments{
\item{nombre_archivo}{}
}
\description{
Tabla de resumen
}
