% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcion Fahrenheit a Celsius.R
\name{convertir_fc}
\alias{convertir_fc}
\title{De fahrenheit a celsius}
\usage{
convertir_fc(temp_fahrenheit)
}
\arguments{
\item{temp_fahrenheit}{Una temperatura en grados fahrenheit (un vector numerico)}
}
\value{
Un vector numérico en grados Celsius
}
\description{
La funcion \code{convertir_fc()} toma una temperatura en grados fahrenheit y la convierte a grados celsius
}
\details{
Estos son detalles extra de la funcion
}
\examples{
convertir_fc(c(0, 32, 100))


}
\seealso{
\code{\link[=convertir_cf]{convertir_cf()}}
}
