% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NH0046.R, R/NH0098.R, R/NH0437.R, R/NH0472.R,
%   R/NH0910.R, R/metadatos.R
\name{get_data}
\alias{get_data}
\title{NH0046}
\format{
Un data frame con estas 35 variables:
\describe{
\item{id}{ID de la estacion, datos categoricos.}
\item{fecha}{Fecha de la medicion, datos temporales.}
\item{temperatura_abrigo_150cm}{Temperatura medida en abrigo a 150 cm, datos numericos.}
\item{temperatura_abrigo_150cm_maxima}{Temperatura maxima en abrigo a 150 cm, datos numericos.}
\item{temperatura_abrigo_150cm_minima}{Temperatura minima en abrigo a 150 cm, datos numericos.}
\item{temperatura_intemperie_5cm_minima}{Temperatura minima medida en intemperie a 5 cm del suelo, valores NA.}
\item{temperatura_intemperie_50cm_minima}{Temperatura minima medida en intemperie a 50 cm del suelo, valores NA.}
\item{temperatura_suelo_5cm_media}{Temperatura media medida en el suelo a 5 cm de profundidad, valores NA.}
\item{temperatura_suelo_10cm_media}{Temperatura media medida en el suelo a 10 cm de profundidad, valores NA.}
\item{temperatura_inte_5cm}{Temperatura medida a 5 cm en intemperie, valores NA.}
\item{temperatura_intemperie_150cm_minima}{Temperatura minima en intemperie a 150 cm, valores NA.}
\item{humedad_suelo}{Humedad del suelo, valores NA.}
\item{precipitacion_pluviometrica}{Precipitacion pluviometrica en mm, datos numericos.}
\item{granizo}{Indica presencia de granizo, valores NA.}
\item{nieve}{Indica presencia de nieve, valores NA.}
\item{heliofania_efectiva}{Heliofania efectiva, valores NA.}
\item{heliofania_relativa}{Heliofania relativa, valores NA.}
\item{tesion_vapor_media}{Tensión de vapor media, valores NA.}
\item{humedad_media}{Humedad relativa media, valores NA.}
\item{humedad_media_8_14_20}{Humedad relativa media registrada a las 8, 14 y 20 horas, valores NA.}
\item{rocio_medio}{Punto de rocio medio, valores NA.}
\item{duracion_follaje_mojado}{Duracion del follaje mojado, valores NA.}
\item{velocidad_viento_200cm_media}{Velocidad media del viento a 200 cm de altura, valores NA.}
\item{direccion_viento_200cm}{Direccion del viento a 200 cm de altura, valores NA.}
\item{velocidad_viento_1000cm_media}{Velocidad media del viento a 1000 cm de altura, valores NA.}
\item{direccion_viento_1000cm}{Direccion del viento a 1000 cm de altura, valores NA.}
\item{velocidad_viento_maxima}{Velocidad maxima del viento registrada, valores NA.}
\item{presion_media}{Presion atmosferica media, valores NA.}
\item{radiacion_global}{Radiacion solar global medida en la estacion, valores NA.}
\item{radiacion_neta}{Radiacion neta medida en la estacion, valores NA.}
\item{evaporacion_tanque}{Evaporacion medida en tanque, valores NA.}
\item{evapotranspiracion_potencial}{Evapotranspiracion potencial, valores NA.}
\item{profundidad_napa}{Profundidad de la napa freatica, valores NA.}
\item{horas_frio}{Cantidad de horas de frio registradas, valores numéricos.}
\item{unidad_frio}{Unidad de medida para las horas de frio, valores NA.}
}

Un data frame con estas 35 variables:
\describe{
\item{id}{ID de la estacion, datos categoricos.}
\item{fecha}{Fecha de la medicion, datos temporales.}
\item{temperatura_abrigo_150cm}{Temperatura medida en abrigo a 150 cm, datos numericos.}
\item{temperatura_abrigo_150cm_maxima}{Temperatura maxima en abrigo a 150 cm, datos numericos.}
\item{temperatura_abrigo_150cm_minima}{Temperatura minima en abrigo a 150 cm, datos numericos.}
\item{temperatura_intemperie_5cm_minima}{Temperatura minima medida en intemperie a 5 cm del suelo, valores NA.}
\item{temperatura_intemperie_50cm_minima}{Temperatura minima medida en intemperie a 50 cm del suelo, valores NA.}
\item{temperatura_suelo_5cm_media}{Temperatura media medida en el suelo a 5 cm de profundidad, valores NA.}
\item{temperatura_suelo_10cm_media}{Temperatura media medida en el suelo a 10 cm de profundidad, valores NA.}
\item{temperatura_inte_5cm}{Temperatura medida a 5 cm en intemperie, valores NA.}
\item{temperatura_intemperie_150cm_minima}{Temperatura minima en intemperie a 150 cm, valores NA.}
\item{humedad_suelo}{Humedad del suelo, valores NA.}
\item{precipitacion_pluviometrica}{Precipitacion pluviometrica en mm, datos numericos.}
\item{granizo}{Indica presencia de granizo, valores NA.}
\item{nieve}{Indica presencia de nieve, valores NA.}
\item{heliofania_efectiva}{Heliofania efectiva, valores NA.}
\item{heliofania_relativa}{Heliofania relativa, valores NA.}
\item{tesion_vapor_media}{Tensión de vapor media, valores NA.}
\item{humedad_media}{Humedad relativa media, valores NA.}
\item{humedad_media_8_14_20}{Humedad relativa media registrada a las 8, 14 y 20 horas, valores NA.}
\item{rocio_medio}{Punto de rocio medio, valores NA.}
\item{duracion_follaje_mojado}{Duracion del follaje mojado, valores NA.}
\item{velocidad_viento_200cm_media}{Velocidad media del viento a 200 cm de altura, valores NA.}
\item{direccion_viento_200cm}{Direccion del viento a 200 cm de altura, valores NA.}
\item{velocidad_viento_1000cm_media}{Velocidad media del viento a 1000 cm de altura, valores NA.}
\item{direccion_viento_1000cm}{Direccion del viento a 1000 cm de altura, valores NA.}
\item{velocidad_viento_maxima}{Velocidad maxima del viento registrada, valores NA.}
\item{presion_media}{Presion atmosferica media, valores NA.}
\item{radiacion_global}{Radiacion solar global medida en la estacion, valores NA.}
\item{radiacion_neta}{Radiacion neta medida en la estacion, valores NA.}
\item{evaporacion_tanque}{Evaporacion medida en tanque, valores NA.}
\item{evapotranspiracion_potencial}{Evapotranspiracion potencial, valores NA.}
\item{profundidad_napa}{Profundidad de la napa freatica, valores NA.}
\item{horas_frio}{Cantidad de horas de frio registradas, valores numéricos.}
\item{unidad_frio}{Unidad de medida para las horas de frio, valores NA.}
}

Un data frame con estas 35 variables:
\describe{
\item{id}{ID de la estacion, datos categoricos.}
\item{fecha}{Fecha de la medicion, datos temporales.}
\item{temperatura_abrigo_150cm}{Temperatura medida en abrigo a 150 cm, datos numericos.}
\item{temperatura_abrigo_150cm_maxima}{Temperatura maxima en abrigo a 150 cm, datos numericos.}
\item{temperatura_abrigo_150cm_minima}{Temperatura minima en abrigo a 150 cm, datos numericos.}
\item{temperatura_intemperie_5cm_minima}{Temperatura minima medida en intemperie a 5 cm del suelo, valores NA.}
\item{temperatura_intemperie_50cm_minima}{Temperatura minima medida en intemperie a 50 cm del suelo, valores NA.}
\item{temperatura_suelo_5cm_media}{Temperatura media medida en el suelo a 5 cm de profundidad, valores NA.}
\item{temperatura_suelo_10cm_media}{Temperatura media medida en el suelo a 10 cm de profundidad, valores NA.}
\item{temperatura_inte_5cm}{Temperatura medida a 5 cm en intemperie, valores NA.}
\item{temperatura_intemperie_150cm_minima}{Temperatura minima en intemperie a 150 cm, valores NA.}
\item{humedad_suelo}{Humedad del suelo, valores NA.}
\item{precipitacion_pluviometrica}{Precipitacion pluviometrica en mm, datos numericos.}
\item{granizo}{Indica presencia de granizo, valores NA.}
\item{nieve}{Indica presencia de nieve, valores NA.}
\item{heliofania_efectiva}{Heliofania efectiva, valores NA.}
\item{heliofania_relativa}{Heliofania relativa, valores NA.}
\item{tesion_vapor_media}{Tensión de vapor media, valores NA.}
\item{humedad_media}{Humedad relativa media, valores NA.}
\item{humedad_media_8_14_20}{Humedad relativa media registrada a las 8, 14 y 20 horas, valores NA.}
\item{rocio_medio}{Punto de rocio medio, valores NA.}
\item{duracion_follaje_mojado}{Duracion del follaje mojado, valores NA.}
\item{velocidad_viento_200cm_media}{Velocidad media del viento a 200 cm de altura, valores NA.}
\item{direccion_viento_200cm}{Direccion del viento a 200 cm de altura, valores NA.}
\item{velocidad_viento_1000cm_media}{Velocidad media del viento a 1000 cm de altura, valores NA.}
\item{direccion_viento_1000cm}{Direccion del viento a 1000 cm de altura, valores NA.}
\item{velocidad_viento_maxima}{Velocidad maxima del viento registrada, valores NA.}
\item{presion_media}{Presion atmosferica media, valores NA.}
\item{radiacion_global}{Radiacion solar global medida en la estacion, valores NA.}
\item{radiacion_neta}{Radiacion neta medida en la estacion, valores NA.}
\item{evaporacion_tanque}{Evaporacion medida en tanque, valores NA.}
\item{evapotranspiracion_potencial}{Evapotranspiracion potencial, valores NA.}
\item{profundidad_napa}{Profundidad de la napa freatica, valores NA.}
\item{horas_frio}{Cantidad de horas de frio registradas, valores numéricos.}
\item{unidad_frio}{Unidad de medida para las horas de frio, valores NA.}
}

Un data frame con estas 35 variables:
\describe{
\item{id}{ID de la estacion, datos categoricos.}
\item{fecha}{Fecha de la medicion, datos temporales.}
\item{temperatura_abrigo_150cm}{Temperatura medida en abrigo a 150 cm, datos numericos.}
\item{temperatura_abrigo_150cm_maxima}{Temperatura maxima en abrigo a 150 cm, datos numericos.}
\item{temperatura_abrigo_150cm_minima}{Temperatura minima en abrigo a 150 cm, datos numericos.}
\item{temperatura_intemperie_5cm_minima}{Temperatura minima medida en intemperie a 5 cm del suelo, valores NA.}
\item{temperatura_intemperie_50cm_minima}{Temperatura minima medida en intemperie a 50 cm del suelo, valores NA.}
\item{temperatura_suelo_5cm_media}{Temperatura media medida en el suelo a 5 cm de profundidad, valores NA.}
\item{temperatura_suelo_10cm_media}{Temperatura media medida en el suelo a 10 cm de profundidad, valores NA.}
\item{temperatura_inte_5cm}{Temperatura medida a 5 cm en intemperie, valores NA.}
\item{temperatura_intemperie_150cm_minima}{Temperatura minima en intemperie a 150 cm, valores NA.}
\item{humedad_suelo}{Humedad del suelo, valores NA.}
\item{precipitacion_pluviometrica}{Precipitacion pluviometrica en mm, datos numericos.}
\item{granizo}{Indica presencia de granizo, valores NA.}
\item{nieve}{Indica presencia de nieve, valores NA.}
\item{heliofania_efectiva}{Heliofania efectiva, valores NA.}
\item{heliofania_relativa}{Heliofania relativa, valores NA.}
\item{tesion_vapor_media}{Tensión de vapor media, valores NA.}
\item{humedad_media}{Humedad relativa media, valores NA.}
\item{humedad_media_8_14_20}{Humedad relativa media registrada a las 8, 14 y 20 horas, valores NA.}
\item{rocio_medio}{Punto de rocio medio, valores NA.}
\item{duracion_follaje_mojado}{Duracion del follaje mojado, valores NA.}
\item{velocidad_viento_200cm_media}{Velocidad media del viento a 200 cm de altura, valores NA.}
\item{direccion_viento_200cm}{Direccion del viento a 200 cm de altura, valores NA.}
\item{velocidad_viento_1000cm_media}{Velocidad media del viento a 1000 cm de altura, valores NA.}
\item{direccion_viento_1000cm}{Direccion del viento a 1000 cm de altura, valores NA.}
\item{velocidad_viento_maxima}{Velocidad maxima del viento registrada, valores NA.}
\item{presion_media}{Presion atmosferica media, valores NA.}
\item{radiacion_global}{Radiacion solar global medida en la estacion, valores NA.}
\item{radiacion_neta}{Radiacion neta medida en la estacion, valores NA.}
\item{evaporacion_tanque}{Evaporacion medida en tanque, valores NA.}
\item{evapotranspiracion_potencial}{Evapotranspiracion potencial, valores NA.}
\item{profundidad_napa}{Profundidad de la napa freatica, valores NA.}
\item{horas_frio}{Cantidad de horas de frio registradas, valores numéricos.}
\item{unidad_frio}{Unidad de medida para las horas de frio, valores NA.}
}

Un data frame con estas 35 variables:
\describe{
\item{id}{ID de la estacion, datos categoricos.}
\item{fecha}{Fecha de la medicion, datos temporales.}
\item{temperatura_abrigo_150cm}{Temperatura medida en abrigo a 150 cm, datos numericos.}
\item{temperatura_abrigo_150cm_maxima}{Temperatura maxima en abrigo a 150 cm, datos numericos.}
\item{temperatura_abrigo_150cm_minima}{Temperatura minima en abrigo a 150 cm, datos numericos.}
\item{temperatura_intemperie_5cm_minima}{Temperatura minima medida en intemperie a 5 cm del suelo, valores NA.}
\item{temperatura_intemperie_50cm_minima}{Temperatura minima medida en intemperie a 50 cm del suelo, valores NA.}
\item{temperatura_suelo_5cm_media}{Temperatura media medida en el suelo a 5 cm de profundidad, valores NA.}
\item{temperatura_suelo_10cm_media}{Temperatura media medida en el suelo a 10 cm de profundidad, valores NA.}
\item{temperatura_inte_5cm}{Temperatura medida a 5 cm en intemperie, valores NA.}
\item{temperatura_intemperie_150cm_minima}{Temperatura minima en intemperie a 150 cm, valores NA.}
\item{humedad_suelo}{Humedad del suelo, valores NA.}
\item{precipitacion_pluviometrica}{Precipitacion pluviometrica en mm, datos numericos.}
\item{granizo}{Indica presencia de granizo, valores NA.}
\item{nieve}{Indica presencia de nieve, valores NA.}
\item{heliofania_efectiva}{Heliofania efectiva, valores NA.}
\item{heliofania_relativa}{Heliofania relativa, valores NA.}
\item{tesion_vapor_media}{Tensión de vapor media, valores NA.}
\item{humedad_media}{Humedad relativa media, valores NA.}
\item{humedad_media_8_14_20}{Humedad relativa media registrada a las 8, 14 y 20 horas, valores NA.}
\item{rocio_medio}{Punto de rocio medio, valores NA.}
\item{duracion_follaje_mojado}{Duracion del follaje mojado, valores NA.}
\item{velocidad_viento_200cm_media}{Velocidad media del viento a 200 cm de altura, valores NA.}
\item{direccion_viento_200cm}{Direccion del viento a 200 cm de altura, valores NA.}
\item{velocidad_viento_1000cm_media}{Velocidad media del viento a 1000 cm de altura, valores NA.}
\item{direccion_viento_1000cm}{Direccion del viento a 1000 cm de altura, valores NA.}
\item{velocidad_viento_maxima}{Velocidad maxima del viento registrada, valores NA.}
\item{presion_media}{Presion atmosferica media, valores NA.}
\item{radiacion_global}{Radiacion solar global medida en la estacion, valores NA.}
\item{radiacion_neta}{Radiacion neta medida en la estacion, valores NA.}
\item{evaporacion_tanque}{Evaporacion medida en tanque, valores NA.}
\item{evapotranspiracion_potencial}{Evapotranspiracion potencial, valores NA.}
\item{profundidad_napa}{Profundidad de la napa freatica, valores NA.}
\item{horas_frio}{Cantidad de horas de frio registradas, valores numéricos.}
\item{unidad_frio}{Unidad de medida para las horas de frio, valores NA.}
}

Un data frame con 258 filas y 11 variables:
\describe{
\item{id}{ID de la estacion, datos categoricos.}
\item{nombre}{Nombre de la estacion, datos categoricos.}
\item{tipo}{Tipo de estacion, posee datos categoricos.}
\item{localidad}{Ciudad de la estacion, datos categoricos.}
\item{provincia}{Provincia de la estacion, datos categoricos.}
\item{lat}{Latitud de la ubicacion de la estacion, datos numericos.}
\item{lon}{Longitud de la ubicacion de la estacion, datos numericos.}
\item{altura}{Altura de la ubicacion de la estacion, datos numericos.}
\item{ubicacion}{Ubicacion exacta de la estacion, datos categoricos.}
\item{desde}{Fecha en la que inicio la estacion, datos temporales.}
\item{hasta}{Fecha en la que finalizo la estacion, datos temporales}
}
}
\source{
Los datos fueron obtenidos de los materiales otorgados en la clase de Programacion II.

Los datos fueron obtenidos de los materiales otorgados en la clase de Programacion II.

Los datos fueron obtenidos de los materiales otorgados en la clase de Programacion II.

Los datos fueron obtenidos de los materiales otorgados en la clase de Programacion II.

Los datos fueron obtenidos de los materiales otorgados en la clase de Programacion II.

Los datos fueron obtenidos de los materiales otorgados en la clase de Programacion II.
}
\usage{
get_data(estacion, destfile)

get_data(estacion, destfile)

get_data(estacion, destfile)

get_data(estacion, destfile)

get_data(estacion, destfile)

get_data(estacion, destfile)
}
\description{
Este es un dataset de ejemplo que contiene informacion sobre estaciones meteorologicas, incluyendo fechas de operacion.

Este es un dataset de ejemplo que contiene informacion sobre estaciones meteorologicas, incluyendo fechas de operacion.

Este es un dataset de ejemplo que contiene informacion sobre estaciones meteorologicas, incluyendo fechas de operacion.

Este es un dataset de ejemplo que contiene informacion sobre estaciones meteorologicas, incluyendo fechas de operacion.

Este es un dataset de ejemplo que contiene informacion sobre estaciones meteorologicas, incluyendo fechas de operacion.

Este es un dataset de ejemplo que contiene informacion sobre estaciones meteorologicas, incluyendo ubicacion, tipo, y fechas de operacion.
}
\examples{
#Cargar el dataset
data(NH0046)

#Ver las primeras filas del dataset
head(NH0046)

#Cargar el dataset
data(NH0098)

#Ver las primeras filas del dataset
head(NH0098)

#Cargar el dataset
data(NH0437)

#Ver las primeras filas del dataset
head(NH0437)

#Cargar el dataset
data(NH0472)

#Ver las primeras filas del dataset
head(NH0472)

#Cargar el dataset
data(NH0910)

#Ver las primeras filas del dataset
head(NH0910)

#Cargar el dataset
data(metadatos)

#Ver las primeras filas del dataset
head(metadatos)

}
