% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tablas.R
\name{tabla_resumen_temperatura}
\alias{tabla_resumen_temperatura}
\title{Tabla de resumen}
\usage{
tabla_resumen_temperatura(nombre_archivo)
}
\arguments{
\item{nombre_archivo}{Un string con el nombre del archivo (sin extensión) que contiene el dataset de datos meteorológicos. Este archivo debe estar ubicado en la carpeta "data" y tener la extensión ".csv.rda".}
}
\value{
Un data frame con cuatro columnas:
\itemize{
\item \code{promedio_temperatura}: la media de la columna \code{temperatura_abrigo_150cm}.
\item \code{desviacion_estandar}: la desviación estándar de la columna \code{temperatura_abrigo_150cm}.
\item \code{minima}: La temperatura minima de la columna \code{temperatura_abrigo_150cm_minima}.
\item \code{maxima}: La temperatura maxima de la columna \code{temperatura_abrigo_150cm_maxima}.
}
}
\description{
La funcion \code{tabla_resumen_temperatura} toma un dataset con informacion sobre estaciones meteorologicas y devuelve un resumen sobre este
}
\details{
Estos son detalles extra de la funcion
}
\examples{
tabla_resumen_temperatura('NH0910')

}
